Replace "rails test" by "rails test:all" and use "rake test:all" if it exist
By mistake, I though that the problem was with the Rake::Task#task_defined?
but actually the problem was in the RAILS test:all command, not the rake one.

The test:all don't accept the -f option for fail-fast, this actually is
passed to the rake test:all command, where it means "the next paramenter will
be the name of the Rakefile" and this was raising an error in the previous
versions.